Move class declarations from __init__.py files into respective modules and import them in __init__.py
